[None][fix] Avoid visual gen sparse config import cycle
