+++ tensorrt_llm/visual_gen/__init__.py
@@ -32,6 +32,8 @@ from .args import (
     CudaGraphConfig,
     ParallelConfig,
     QuantAttentionConfig,
+    SkipSoftmaxConfig,
+    SparseAttentionConfig,
     TeaCacheConfig,
     TorchCompileConfig,
     VisualGenArgs,
@@ -56,6 +58,8 @@ __all__ = [
     "ParallelConfig",
     "AttentionConfig",
     "QuantAttentionConfig",
+    "SparseAttentionConfig",
+    "SkipSoftmaxConfig",
     "CacheConfig",
     "TeaCacheConfig",
     "CacheDiTConfig",
